Restructuring complete. create_db is rewritten
