+++ VDB/VDB.vcxproj
@@ -127,6 +127,8 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="main.cpp" />
+    <ClCompile Include="vdb_console.cpp" />
+    <ClCompile Include="vdb_query_proc.cpp" />
     <ClCompile Include="vdb_response.cpp" />
     <ClCompile Include="vdb_row.cpp" />
     <ClCompile Include="vdb_table.cpp" />
@@ -140,6 +142,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="vdb_api.h" />
+    <ClInclude Include="vdb_console.h" />
     <ClInclude Include="vdb_query_proc.h" />
     <ClInclude Include="vdb_response.h" />
     <ClInclude Include="vdb_row.h" />
