+++ VDB/vdb_console.h
@@ -1,0 +1,13 @@
+#ifndef VDB_CONSOLE_H
+#define VDB_CONSOLE_H
+
+#include <string>
+
+namespace vdb
+{
+	int query_cout(std::string query);
+	void vdbms_cout();
+
+}
+
+#endif // !VDB_CONSOLE_H
