+++ VDB/vdb_table.cpp
@@ -1,586 +1,139 @@
 #include "vdb_table.h"
+#include "vdb_query_proc.h"
 #include "vdb_utils.h"
 
-namespace 
-{ 
-	// This is query processor stuff --------------------------> vvvvvvvvvvvvvvvvvvvvvvvvvvv
-	struct Node
-	{
-		std::variant<vdb::Value, uint8_t, char> value;
-		struct Node *right;
-		struct Node *left;
-		Node()
+#include <sstream>
+#include <cctype>
+
+struct vdb::Table::column
+{
+	char name[32];
+	uint8_t type;
+	uint8_t size;
+};
+
+namespace
+{
+	void get_name(std::stringstream &ss, std::string &to)
+	{ 
+		// skip everything till `
+		while (ss.peek() != '`')
+			ss.get();
+		ss.get();
+
+		while (ss.peek() != '`')
 		{
-			right = left = nullptr;
+			if (ss.peek() == '\\') // skip escaped char
+				ss.get();
+			else
+				to += ss.get();
 		}
-	};
-	bool is_match(vdb::Row &row, const Node *tree)
+		ss.get(); // skip last `
+		ss.peek(); // to turn on eof bit (if it's the end)
+	}
+
+	size_t col_type(std::string &name)
 	{
-		if (tree->value.index() == 0 || tree->value.index() == 1)
-			return false;
-		if (tree->value.index() == 2)
-		{
-			if ((std::get<char>(tree->value) != '&') && (std::get<char>(tree->value) != '|'))
-			{
-				if ((tree->left->value.index() == 0 ? std::get<uint8_t>(tree->right->value) : std::get<uint8_t>(tree->left->value)) == 0xFF)
-					return false; // return false if there's no column with such name (look set_tree()), 0xFF is indicator.
-
-				switch (std::get<char>(tree->value))
-				{
-					case '=':
-					{
-						switch (row[tree->left->value.index() == 0 ? std::get<uint8_t>(tree->right->value) : std::get<uint8_t>(tree->left->value)].get_type())
-						{
-							case 0:
-							{
-								return (tree->left->value.index() == 0 ? static_cast<int>(std::get<vdb::Value>(tree->left->value)) : static_cast<int>(row[std::get<uint8_t>(tree->left->value)])) == (tree->right->value.index() == 0 ? static_cast<int>(std::get<vdb::Value>(tree->right->value)) : static_cast<int>(row[std::get<uint8_t>(tree->right->value)]));
-							}; break;
-							case 1:
-							{
-								return (tree->left->value.index() == 0 ? static_cast<double>(std::get<vdb::Value>(tree->left->value)) : static_cast<double>(row[std::get<uint8_t>(tree->left->value)])) == (tree->right->value.index() == 0 ? static_cast<double>(std::get<vdb::Value>(tree->right->value)) : static_cast<double>(row[std::get<uint8_t>(tree->right->value)]));
-							}; break;
-							case 2:
-							{
-								return (tree->left->value.index() == 0 ? static_cast<char>(std::get<vdb::Value>(tree->left->value)) : static_cast<char>(row[std::get<uint8_t>(tree->left->value)])) == (tree->right->value.index() == 0 ? static_cast<char>(std::get<vdb::Value>(tree->right->value)) : static_cast<char>(row[std::get<uint8_t>(tree->right->value)]));
-							}; break;
-							case 3:
-							case 4:
-							{
-								// strcmp() returns 0 if strings are equals.
-								return !static_cast<bool>(std::strcmp((tree->left->value.index() == 0 ? static_cast<char *>(std::get<vdb::Value>(tree->left->value)) : static_cast<char *>(row[std::get<uint8_t>(tree->left->value)])), (tree->right->value.index() == 0 ? static_cast<char *>(std::get<vdb::Value>(tree->right->value)) : static_cast<char *>(row[std::get<uint8_t>(tree->right->value)]))));
-							}; break;
-							default:
-							{
-								// Maybe I should write something here?
-							}
-						}
-					}; break;
-					case '!':
-					{
-						switch (row[tree->left->value.index() == 0 ? std::get<uint8_t>(tree->right->value) : std::get<uint8_t>(tree->left->value)].get_type())
-						{
-							case 0:
-							{
-								return (tree->left->value.index() == 0 ? static_cast<int>(std::get<vdb::Value>(tree->left->value)) : static_cast<int>(row[std::get<uint8_t>(tree->left->value)])) != (tree->right->value.index() == 0 ? static_cast<int>(std::get<vdb::Value>(tree->right->value)) : static_cast<int>(row[std::get<uint8_t>(tree->right->value)]));
-							}; break;
-							case 1:
-							{
-								return (tree->left->value.index() == 0 ? static_cast<double>(std::get<vdb::Value>(tree->left->value)) : static_cast<double>(row[std::get<uint8_t>(tree->left->value)])) != (tree->right->value.index() == 0 ? static_cast<double>(std::get<vdb::Value>(tree->right->value)) : static_cast<double>(row[std::get<uint8_t>(tree->right->value)]));
-							}; break;
-							case 2:
-							{
-								return (tree->left->value.index() == 0 ? static_cast<char>(std::get<vdb::Value>(tree->left->value)) : static_cast<char>(row[std::get<uint8_t>(tree->left->value)])) != (tree->right->value.index() == 0 ? static_cast<char>(std::get<vdb::Value>(tree->right->value)) : static_cast<char>(row[std::get<uint8_t>(tree->right->value)]));
-							}; break;
-							case 3:
-							case 4:
-							{
-								// strcmp() returns 0 if strings are equals.
-								return static_cast<bool>(std::strcmp((tree->left->value.index() == 0 ? static_cast<char *>(std::get<vdb::Value>(tree->left->value)) : static_cast<char *>(row[std::get<uint8_t>(tree->left->value)])), (tree->right->value.index() == 0 ? static_cast<char *>(std::get<vdb::Value>(tree->right->value)) : static_cast<char *>(row[std::get<uint8_t>(tree->right->value)]))));
-							}; break;
-							default:
-							{
-								// Maybe I should write something here?
-							}
-						}
-					}; break;
-					case '>':
-					{
-						switch (row[tree->left->value.index() == 0 ? std::get<uint8_t>(tree->right->value) : std::get<uint8_t>(tree->left->value)].get_type())
-						{
-							case 0:
-							{
-								return (tree->left->value.index() == 0 ? static_cast<int>(std::get<vdb::Value>(tree->left->value)) : static_cast<int>(row[std::get<uint8_t>(tree->left->value)])) > (tree->right->value.index() == 0 ? static_cast<int>(std::get<vdb::Value>(tree->right->value)) : static_cast<int>(row[std::get<uint8_t>(tree->right->value)]));
-							}; break;
-							case 1:
-							{
-								return (tree->left->value.index() == 0 ? static_cast<double>(std::get<vdb::Value>(tree->left->value)) : static_cast<double>(row[std::get<uint8_t>(tree->left->value)])) > (tree->right->value.index() == 0 ? static_cast<double>(std::get<vdb::Value>(tree->right->value)) : static_cast<double>(row[std::get<uint8_t>(tree->right->value)]));
-							}; break;
-							default:
-							{
-								// Maybe I should write something here?
-							}
-						}
-					}; break;
-					case '<':
-					{
-						switch (row[tree->left->value.index() == 0 ? std::get<uint8_t>(tree->right->value) : std::get<uint8_t>(tree->left->value)].get_type())
-						{
-							case 0:
-							{
-								return (tree->left->value.index() == 0 ? static_cast<int>(std::get<vdb::Value>(tree->left->value)) : static_cast<int>(row[std::get<uint8_t>(tree->left->value)])) < (tree->right->value.index() == 0 ? static_cast<int>(std::get<vdb::Value>(tree->right->value)) : static_cast<int>(row[std::get<uint8_t>(tree->right->value)]));
-							}; break;
-							case 1:
-							{
-								return (tree->left->value.index() == 0 ? static_cast<double>(std::get<vdb::Value>(tree->left->value)) : static_cast<double>(row[std::get<uint8_t>(tree->left->value)])) < (tree->right->value.index() == 0 ? static_cast<double>(std::get<vdb::Value>(tree->right->value)) : static_cast<double>(row[std::get<uint8_t>(tree->right->value)]));
-							}; break;
-							default:
-							{
-								// Maybe I should write something here?
-							}
-						}
-					}; break;
-					case 'g': // >=
-					{
-						switch (row[tree->left->value.index() == 0 ? std::get<uint8_t>(tree->right->value) : std::get<uint8_t>(tree->left->value)].get_type())
-						{
-							case 0:
-							{
-								return (tree->left->value.index() == 0 ? static_cast<int>(std::get<vdb::Value>(tree->left->value)) : static_cast<int>(row[std::get<uint8_t>(tree->left->value)])) >= (tree->right->value.index() == 0 ? static_cast<int>(std::get<vdb::Value>(tree->right->value)) : static_cast<int>(row[std::get<uint8_t>(tree->right->value)]));
-							}; break;
-							case 1:
-							{
-								return (tree->left->value.index() == 0 ? static_cast<double>(std::get<vdb::Value>(tree->left->value)) : static_cast<double>(row[std::get<uint8_t>(tree->left->value)])) >= (tree->right->value.index() == 0 ? static_cast<double>(std::get<vdb::Value>(tree->right->value)) : static_cast<double>(row[std::get<uint8_t>(tree->right->value)]));
-							}; break;
-							default:
-							{
-								// Maybe I should write something here?
-							}
-						}
-					}; break;
-					case 'l': // <=
-					{
-						switch (row[tree->left->value.index() == 0 ? std::get<uint8_t>(tree->right->value) : std::get<uint8_t>(tree->left->value)].get_type())
-						{
-							case 0:
-							{
-								return (tree->left->value.index() == 0 ? static_cast<int>(std::get<vdb::Value>(tree->left->value)) : static_cast<int>(row[std::get<uint8_t>(tree->left->value)])) <= (tree->right->value.index() == 0 ? static_cast<int>(std::get<vdb::Value>(tree->right->value)) : static_cast<int>(row[std::get<uint8_t>(tree->right->value)]));
-							}; break;
-							case 1:
-							{
-								return (tree->left->value.index() == 0 ? static_cast<double>(std::get<vdb::Value>(tree->left->value)) : static_cast<double>(row[std::get<uint8_t>(tree->left->value)])) <= (tree->right->value.index() == 0 ? static_cast<double>(std::get<vdb::Value>(tree->right->value)) : static_cast<double>(row[std::get<uint8_t>(tree->right->value)]));
-							}; break;
-							default:
-							{
-								// Maybe I should write something here?
-							}
-						}
-					}; break;
-					default:
-						break;
-				}
-			}
-		}
-
-		bool op_1 = is_match(row, tree->left);
-		bool op_2 = is_match(row, tree->right);
-
-		if (std::get<char>(tree->value) == '|')
-			return op_1 || op_2;
+		if (name == "int")
+			return 0;
+		else if (name == "char")
+			return 2;
+		else if (name == "str32")
+			return 3;
+		else if (name == "str64")
+			return 4;
 		else
-			return op_1 && op_2;
-
-	}
-	void destroy_tree(struct Node *node)
-	{
-		if (node != nullptr)
-		{
-			destroy_tree(node->left);
-			destroy_tree(node->right);
-			delete node;
-		}
-	}
-	//TODO INSIDE
-	void set_tree(std::string str, Node *tree, const vdb::Table &table)
-	{
-		uint8_t nesting_level = 0;
-		bool operations_left = false;
-
-		// To remove wrapping brackets, for example (a > b), crucial for function to work
-		// It should also distingiush something like (a == b) && (c < d) from wrapping brackets
-		// TODO: There're problems with 3 or above levels of nesting <-------------------------------------------------
-		if (str[0] == '(' && str[str.length() - 1] == ')')
-		{
-			// If there's ')' braket, but '(' braket (open = false) doesn't accured, no action should be done
-			bool open = false, error = false;
-			// skip first and last character
-			for (size_t i = 1; i < str.length() - 1; ++i)
-			{
-				if (str[i] == '(')
-				{
-					open = true;
-				}
-				else if (str[i] == ')')
-				{
-					if (!open)
-					{
-						error = true;
-						break;
-					}
-					else
-						open = false;
-				}
-			}
-
-			if (!error)
-				str = str.substr(1, str.length() - 2);
-		}
-
-
-		for (size_t i = 0; i < str.length(); ++i)
-		{
-			if (str[i] == '(')
-			{
-				++nesting_level;
-			}
-			else if (str[i] == ')')
-			{
-				--nesting_level;
-			}
-			else if (str[i] == '`')
-			{
-				while (str[++i] != '`')
-					if (str[i] == '\\' && (i + 1) < str.length() && (i + 2) != str.length())
-						++i;
-			}
-			else if (str[i] == '\"')
-			{
-				while (str[++i] != '\"')
-					if (str[i] == '\\' && (i + 1) < str.length() && (i + 2) != str.length())
-						++i;
-			}
-			else if ((str[i] == '=' && str[i + 1] == '=') || (str[i] == '!' && str[i + 1] == '=') || str[i] == '>' || str[i] == '<' || (str[i] == '>' && str[i + 1] == '=') || (str[i] == '<' && str[i + 1] == '='))
-			{
-				operations_left = true;
-			}
-			else if ((str[i] == '&' || str[i] == '|') && nesting_level == 0)
-			{
-				tree->value = str[i];
-				tree->left = new Node();
-				tree->right = new Node();
-				set_tree(str.substr(0, i), tree->left, table);
-				set_tree(str.substr(i + 2, str.length() - i - 2), tree->right, table);
-				return;
-			}
-		}
-
-		nesting_level = 0;
-
-		// If there's no &&, || operations.
-		if (operations_left)
-		{
-			for (size_t i = 0; i < str.length(); ++i)
-			{
-				if (str[i] == '(')
-				{
-					++nesting_level;
-				}
-				else if (str[i] == ')')
-				{
-					--nesting_level;
-				}
-				else if (str[i] == '`')
-				{
-					while (str[++i] != '`')
-						if (str[i] == '\\' && (i + 1) < str.length() && (i + 2) != str.length())
-							++i;
-				}
-				else if (str[i] == '\"')
-				{
-					while (str[++i] != '\"')
-						if (str[i] == '\\' && (i + 1) < str.length() && (i + 2) != str.length())
-							++i;
-				}
-				else if (((str[i] == '=' && str[i + 1] == '=') || (str[i] == '!' && str[i + 1] == '=') || str[i] == '>' || str[i] == '<' || (str[i] == '>' && str[i + 1] == '=') || (str[i] == '<' && str[i + 1] == '=')) && nesting_level == 0)
-				{
-					uint8_t is_double_char = 0; // to get correct substring when operator is two character long (<= and >=)
-					if (str[i] == '>' && str[i + 1] == '=')
-					{
-						tree->value = 'g';
-						is_double_char = 1;
-					}
-					else if (str[i] == '<' && str[i + 1] == '=')
-					{
-						tree->value = 'l';
-						is_double_char = 1;
-					}
-					else if (str[i] == '!' && str[i + 1] == '=')
-					{
-						tree->value = '!';
-						is_double_char = 1;
-					}
-					else if (str[i] == '=' && str[i + 1] == '=')
-					{
-						tree->value = '=';
-						is_double_char = 1;
-					}
-					else
-						tree->value = str[i];
-
-					tree->left = new Node();
-					tree->right = new Node();
-					set_tree(str.substr(0, i), tree->left, table);
-					set_tree(str.substr(i + 1 + is_double_char, str.length() - i - 1 - is_double_char), tree->right, table);
-					return;
-				}
-			}
-		}
-		else
-		{
-			if (str[0] == '`') // column name
-			{
-				str = str.substr(1, str.length() - 2);
-
-				unescape(str);
-
-				for (uint8_t i = 0; i < table.get_colcount(); i++)
-				{
-					if (table.get_col_name(i) == str)
-					{
-						tree->value = i;
-						return;
-					}
-				}
-
-				tree->value = uint8_t(0xFF); // To indicate that there's no column with such name
-			}
-			else if (str[0] == '\'') // char value
-			{
-				vdb::Value val(str[1]);
-				tree->value = val;
-			}
-			else if (str[0] == '\"') // string value
-			{
-				str = str.substr(1, str.length() - 2);
-				unescape(str);
-				vdb::Value val(str.c_str());
-				tree->value = val;
-			}
-			else // int or double value
-			{
-				try
-				{
-					vdb::Value val(std::stod(str));
-					tree->value = val;
-				}
-				catch (const std::exception &)
-				{
-					std::cout << "I TRUSTED YOU!!!!!" << std::endl;
-					vdb::Value val(0);
-					tree->value = val;
-				}
-
-			}
-		}
-	}
-	// This is query processor stuff --------------------------> ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-
-	void print_response(vdb::Response &response)
-	{
-		for (size_t i = 0; i < response.size(); i++)
-		{
-			for (size_t j = 0; j < response[i].size(); j++)
-			{
-				std::cout << response[i][j].to_string() << "\t";
-			}
-			std::cout << std::endl;
-		}
-	}
-	// Should I delete print_rows?
-	void print_rows(vdb::Row *row, size_t size)
-	{
-		for (size_t i = 0; i < size; ++i)
-		{
-			for (size_t j = 0; j < row[i].size(); ++j)
-			{
-				std::cout << row[i][j].to_string() << " ";
-			}
-			std::cout << std::endl;
-		}
-	}
-	void print_table_colnames(const vdb::Table table)
-	{
-		for (size_t i = 0; i < table.get_colcount(); ++i)
-		{
-			std::cout << table.get_col_name(i) << "\t";
-		}
-		std::cout << std::endl;
+			return 1; // if double or anything else
 	}
 }
 
-bool vdb::create_db(std::string desc) //check if desc has correct syntax
+bool vdb::create_db(std::string desc)
 {
-	std::string file_name;
-	std::fstream file;
+	/*
+		Syntax correctness is caller's duty.
+		Meta layout: [meta_size][colcount][rowsize][rowcount][[coltype][colname]...]
+	*/
+	trim(desc);
+	std::stringstream ss{desc};
+	std::string filename; // use it as filename
 
-	remove_spaces(desc);
+	get_name(ss, filename); // `name with possible \` char`
+	unescape(filename);
+	filename.append(".vdb");
 
-	// pos1 ... pos2
-	size_t pos1;
-	size_t pos2 = 0;
-	while (desc[++pos2] != '`') // to skip all escaped `
-	{
-		if (desc[pos2] == '\\')
-			++pos2;
-	}
-
-	// db path with name
-	file_name = desc.substr(1, pos2 - 1);
-
-	unescape(file_name); // \` -> `, \\ -> \
-
-	if (file_name.substr(file_name.length() - 4, 4) != ".vdb")
-		file_name.append(".vdb");
-
-	file.open(file_name, std::ios::binary | std::ios::out);
+	std::fstream file(filename, std::ios::binary | std::ios::out);
 
 	if (!file.is_open())
 		return false;
 
-	uint16_t two_bytes;
-	pos1 = pos2;
-	pos2 = desc.find_first_of(' ', pos1 + 2);
-	uint8_t colcount = std::stoi(desc.substr(pos1 + 2, pos2 - pos1));
+	uint16_t meta_size, colcount = 0, rowsize = 0;
 
-	//meta size
-	two_bytes = 7 + (33 * colcount);
-
-	file.write((char *)&two_bytes, 2);
-	file.write((char *)&colcount, 1);
-
-	two_bytes = 0;
-	// number of rows (= 0)
-	file.write((char *)&two_bytes, 2);
+	file.write((char *)&rowsize, 8); // skip this for now (will be rewritten)
 
 	int colsize[] = {4, 8, 1, 32, 64};
 
-	uint16_t rowsize = 0;
+	std::string colinfo; 
+	uint8_t coltype;
 
-	std::string name;
-
-	for (int i = 0, type; i < colcount; i++)
+	while (!ss.eof() && ss) // while no troubles occured
 	{
-		pos1 = pos2;
-		pos2 = desc.find_first_of(' ', pos1 + 1);
-		type = std::stoi(desc.substr(pos1 + 1, pos2 - pos1));
-		rowsize += colsize[type];
-		file.write((char *)&type, 1);
-		pos1 = pos2 + 2;
-		pos2 = pos1 - 1;
-		//pos2 = desc.find_first_of('`', pos1);
-		while (desc[++pos2] != '`') // to skip all escaped `
-		{
-			if (desc[pos2] == '\\')
-				++pos2;
-		}
-		name = desc.substr(pos1, pos2 - pos1);
-		unescape(name);
-		file.write(name.c_str(), 32);
-		++pos2;
+		while (!isalpha(ss.peek()))
+			ss.get();
+
+		ss >> colinfo; // get literal coltype
+		coltype = col_type(colinfo);  
+
+		colinfo.clear();
+		get_name(ss, colinfo); // get colname
+		unescape(colinfo);
+
+		file.write((char *)&coltype, 1);
+		file.write(colinfo.c_str(), 32);
+
+		++colcount;
+		rowsize += colsize[coltype];
 	}
 
-	// write rowsize
+	meta_size = 8 + (33 * colcount);
+
+	file.seekp(0);
+
+	file.write((char *)&meta_size, 2);
+	file.write((char *)&colcount, 2);
+	file.write((char *)&rowsize, 2);
+	
+	rowsize = 0; //use it as rowcount (= 0)
 	file.write((char *)&rowsize, 2);
 
 	file.close();
-
 	return true;
 }
 
-int vdb::query_cout(std::string query)
+//todo DEBUG --- start ---
+#include <iostream>
+void vdb::Table::print_meta()
 {
+	std::cout << "Meta size: \t" << meta_size << std::endl;
+	std::cout << "Column count: \t" << int(colcount) << std::endl;
+	std::cout << "Row count: \t" << rowcount << std::endl;
+	std::cout << "Row size: \t" << rowsize << std::endl << std::endl;
 
-	ltrim(query);
-
-	if (query[0] != '`')
+	for (int i = 0, type; i < colcount; i++)
 	{
-		std::string substr = next_token(query, " ");
-
-		if (substr == "create")
-		{
-			substr = next_token(query, " ");
-
-			if (substr == "table")
-			{
-				vdb::create_db(query);
-				return 0;
-			}
-			else
-				return 2;
-		}
-		else if (substr == "cls")
-		{
-			system("cls");
-			return 0;
-		}
-		else if (substr == "exit")
-			return 1812;
-		else
-			return 2;
+		type = cols[i].type;
+		std::cout << "#" << i + 1 << " Column name: " << cols[i].name << std::endl;
+		std::cout << "#" << i + 1 << " Column type: " << (type == 0 ? "int" : (type == 1 ? "double" : (type == 2 ? "char" : (type == 3 ? "STR32" : "STR64")))) << std::endl;
 	}
 
-	size_t i = 0;
-	while (query[++i] != '`')
-		if (query[i] == '\\' && (i + 1) < query.length() && (i + 2) != query.length())
-			++i;
-
-	std::string table_name = query.substr(1, i - 1);
-
-	query = query.substr(i + 2);
-
-	unescape(table_name);
-
-	vdb::Table table;
-	table.open(table_name);
-
-	if (!table.is_open())
-		return -1;
-
-	std::string substr = next_token(query, " ");
-
-	if (substr == "select_all")
-	{
-		vdb::Response resp = table.select_all();
-		print_table_colnames(table);
-		print_response(resp);
-	}
-	else if (substr == "insert_into")
-	{
-		table.insert_into(query);
-	}
-	else if (substr == "select_where")
-	{
-		vdb::Response resp = table.select_where(query);
-		print_table_colnames(table);
-		print_response(resp);
-	}
-	else if (substr == "clear")
-	{
-		table.clear();
-	}
-	else if (substr == "remove_line")
-	{
-		try
-		{
-			size_t line = std::stoi(query);
-			table.remove_line(line);
-		}
-		catch (const std::exception &)
-		{
-			std::cout << "Fuck you, asshole!";
-		}
-	}
-	else if (substr == "delete")
-	{
-		table.close();
-		std::string query("del ");
-		query.append(table_name);
-		if (table_name.substr(table_name.length() - 4, 4) != ".vdb")
-			query.append(".vdb");
-		std::system(query.c_str());
-		return 0;
-	}
-
-	table.close();
-	return 0;
+	std::cout << std::endl;
 }
+//todo DEBUG --- end ---
 
 
 
 // Class managment
-vdb::Table::Table()
-{
-	opened = false;
-}
+vdb::Table::Table() : opened{false} { }
+
 vdb::Table::Table(const Table &table)
 {
 	if (!table.opened)
@@ -594,7 +147,7 @@ vdb::Table::Table(const Table &table)
 		rowcount = table.rowcount;
 		rowsize = table.rowsize;
 
-		cols = new vdb::column[colcount];
+		cols = new column[colcount];
 
 		for (size_t i = 0; i < colcount; ++i)
 		{
@@ -628,7 +181,7 @@ vdb::Table &vdb::Table::operator=(const vdb::Table &table)
 		rowcount = table.rowcount;
 		rowsize = table.rowsize;
 
-		cols = new vdb::column[colcount];
+		cols = new column[colcount];
 
 		for (size_t i = 0; i < colcount; ++i)
 		{
@@ -672,10 +225,12 @@ bool vdb::Table::open(const std::string &name)
 		return false;
 
 	file.read((char *)&meta_size, 2);
-	file.read((char *)&colcount, 1);
+	file.read((char *)&colcount, 2);
+	file.read((char *)&rowsize, 2);
 	file.read((char *)&rowcount, 2);
+	
 
-	cols = new vdb::column[colcount];
+	cols = new column[colcount];
 
 	for (int i = 0; i < colcount; i++)
 	{
@@ -684,8 +239,7 @@ bool vdb::Table::open(const std::string &name)
 		file.read((char *)&(cols[i].name), 32);
 	}
 
-	file.read((char *)&rowsize, 2);
-
+	
 	file.seekg(0);
 
 	opened = true;
@@ -935,8 +489,10 @@ vdb::Response vdb::Table::select_all()
 	return resp;
 }
 // TODO INSIDE!!!
-vdb::Response vdb::Table::select_where(std::string &condition)
+vdb::Response vdb::Table::select_where(std::string condition)
 {
+	using namespace vdb_impl;
+
 	// All excess blanks must be truncated
 	ltrim(condition); // without it mechanism crashes because iterator can't be decremented under the condition.begin()
 	for (auto it = condition.begin(); it != condition.end(); ++it) // TODO: The leading blank will fuck my algorithm
@@ -981,11 +537,6 @@ vdb::Response vdb::Table::select_where(std::string &condition)
 	return resp;
 
 	return resp;
-}
-vdb::Response vdb::Table::select_where(const char *condition)
-{
-	std::string cond(condition);
-	return select_where(cond);
 }
